修正脚本

+++ project/tools/getMd5List.js
@@ -1,5 +1,6 @@
 const crypto = require('crypto');
 const fs = require('fs');
+const fetch = require('node-fetch');
 const mapValues = require('lodash/mapValues');
 
 function getFileMd5(filename) {
@@ -25,20 +26,20 @@ async function getMd5List(path, list) {
     }
     return list;
 }
-async function getMapList(platform, distpath) {
+async function getMapList(platform, remote_md5_url) {
     let app = {};
     try {
-        app = JSON.parse(fs.readFileSync(distpath+'/'+platform+'_md5.json'));
+        app = await fetch(remote_md5_url, {method: 'get'}).then((response) => response.json());
     } catch(e) {}
     const www = await getMd5List('./www', {});
     fs.writeFileSync('./'+platform+'_md5.json', JSON.stringify(www));
     return mapValues(www, (v, k)=>v===app[k]);
 }
-async function genNeedDownloadFiles(platform, distpath) {
-    const list = await getMapList(platform, distpath);
+async function genNeedDownloadFiles(platform, remote_md5_url) {
+    const list = await getMapList(platform, remote_md5_url);
     const needCopyFiles = [];
     for (const key in list) {
-        if (list[key]) {
+        if (list[key] && !/index.(ios|android).bundle/.test(key)) {
             needCopyFiles.push(key.replace(/^\.\/www\//, ''));
             console.log("needCopyFiles file:", key);
             fs.unlinkSync(key);
